+++ crates/holochain_integrity_types/src/chain.rs
@@ -16,50 +16,53 @@ mod test;
 /// Filter source chain items.
 /// Starting from some chain position given as an [`ActionHash`]
 /// the chain is walked backwards to genesis.
-/// The filter can stop early by specifying the number of
-/// chain items to take and / or an [`ActionHash`] to consume until.
+/// The filter can stop early by specifying stop conditions:
+/// A maximum number of items is reached, a given [`ActionHash`] is found, or until
+/// a given timestamp has been reached.
+/// Multiple `[`ActionHash`] can be provided. The filter will stop at and take the first one found.
+/// When providing a Timestamp, the filter will stop at and take the last item that has a
+/// Timestamp younger than the provided one.
 pub struct ChainFilter<H: Eq + Ord + std::hash::Hash = ActionHash> {
     /// The starting position of the filter.
     pub chain_top: H,
-    /// The filters that have been applied.
-    /// Defaults to [`ChainFilters::ToGenesis`].
-    pub filters: ChainFilters<H>,
+    /// The stop conditions used by this filter.
+    /// Defaults to [`StopConditions::ToGenesis`].
+    pub stop_conditions: StopConditions<H>,
     /// Should the query return any entries that are
     /// cached at the agent activity to save network hops.
     pub include_cached_entries: bool,
 }
 
 #[derive(Serialize, Deserialize, Debug, Eq, Clone)]
-/// Specify which [`Action`](crate::action::Action)s to allow through
-/// this filter.
-pub enum ChainFilters<H: Eq + Ord + std::hash::Hash = ActionHash> {
+/// Specify when to stop walking down the chain.
+pub enum StopConditions<H: Eq + Ord + std::hash::Hash = ActionHash> {
     /// Allow all up to genesis.
     ToGenesis,
-    /// Take this many (inclusive of the starting position).
+    /// Take this many items (inclusive of the starting position).
     Take(u32),
-    /// Continue until timestamp is reached
+    /// Continue until some timestamp is passed.
     UntilTimestamp(Timestamp),
     /// Continue until one of these hashes is found.
     UntilHash(HashSet<H>),
-    /// Combination of take, until_hash and until_timestamp.
+    /// Combination of Take, UntilTimestamp and UntilHash.
     /// Whichever is the smaller set.
     Multiple(u32, HashSet<H>, Timestamp),
 }
 
-/// Create a deterministic hash to compare filters.
-impl<H: Eq + Ord + std::hash::Hash> core::hash::Hash for ChainFilters<H> {
+/// Create a deterministic hash to compare StopConditions.
+impl<H: Eq + Ord + std::hash::Hash> core::hash::Hash for StopConditions<H> {
     fn hash<HH: std::hash::Hasher>(&self, state: &mut HH) {
         core::mem::discriminant(self).hash(state);
         match self {
-            ChainFilters::ToGenesis => (),
-            ChainFilters::Take(t) => t.hash(state),
-            ChainFilters::UntilTimestamp(ts) => ts.hash(state),
-            ChainFilters::UntilHash(u) => {
+            StopConditions::ToGenesis => (),
+            StopConditions::Take(t) => t.hash(state),
+            StopConditions::UntilTimestamp(ts) => ts.hash(state),
+            StopConditions::UntilHash(u) => {
                 let mut u: Vec<_> = u.iter().collect();
                 u.sort_unstable();
                 u.hash(state);
             }
-            ChainFilters::Multiple(n, u, t) => {
+            StopConditions::Multiple(n, u, t) => {
                 let mut u: Vec<_> = u.iter().collect();
                 u.sort_unstable();
                 u.hash(state);
@@ -70,8 +73,8 @@ impl<H: Eq + Ord + std::hash::Hash> core::hash::Hash for ChainFilters<H> {
     }
 }
 
-/// Implement a deterministic partial eq to compare ChainFilters.
-impl<H: Eq + Ord + std::hash::Hash> core::cmp::PartialEq for ChainFilters<H> {
+/// Implement a deterministic partial eq to compare StopConditions.
+impl<H: Eq + Ord + std::hash::Hash> core::cmp::PartialEq for StopConditions<H> {
     fn eq(&self, other: &Self) -> bool {
         match (self, other) {
             (Self::Take(l0), Self::Take(r0)) => l0 == r0,
@@ -96,12 +99,12 @@ impl<H: Eq + Ord + std::hash::Hash> core::cmp::PartialEq for ChainFilters<H> {
 }
 
 #[derive(Clone, Debug, Serialize, Deserialize, PartialEq, Eq)]
-/// Input to the `must_get_agent_activity` call.
+/// Input of the `must_get_agent_activity` call.
 pub struct MustGetAgentActivityInput {
     /// The author of the chain that you are requesting
     /// activity from.
     pub author: AgentPubKey,
-    /// The filter on the chains activity.
+    /// The filter on the chain's activity.
     pub chain_filter: ChainFilter,
 }
 
@@ -112,7 +115,7 @@ impl<H: Eq + Ord + std::hash::Hash> ChainFilter<H> {
     pub fn new(chain_top: H) -> Self {
         Self {
             chain_top,
-            filters: Default::default(),
+            stop_conditions: Default::default(),
             include_cached_entries: false,
         }
     }
@@ -120,12 +123,12 @@ impl<H: Eq + Ord + std::hash::Hash> ChainFilter<H> {
     /// Take up to `n` actions including the starting position.
     /// This may return less then `n` actions.
     pub fn take(mut self, n: u32) -> Self {
-        self.filters = match self.filters {
-            ChainFilters::ToGenesis => ChainFilters::Take(n),
-            ChainFilters::Take(old_n) => ChainFilters::Take(old_n.min(n)),
-            ChainFilters::UntilTimestamp(t) => ChainFilters::Multiple(n, HashSet::new(), t),
-            ChainFilters::UntilHash(u) => ChainFilters::Multiple(n, u, Timestamp(0)),
-            ChainFilters::Multiple(old_n, u, t) => ChainFilters::Multiple(old_n.min(n), u, t),
+        self.stop_conditions = match self.stop_conditions {
+            StopConditions::ToGenesis => StopConditions::Take(n),
+            StopConditions::Take(old_n) => StopConditions::Take(old_n.min(n)),
+            StopConditions::UntilTimestamp(t) => StopConditions::Multiple(n, HashSet::new(), t),
+            StopConditions::UntilHash(u) => StopConditions::Multiple(n, u, Timestamp(0)),
+            StopConditions::Multiple(old_n, u, t) => StopConditions::Multiple(old_n.min(n), u, t),
         };
         self
     }
@@ -139,12 +142,14 @@ impl<H: Eq + Ord + std::hash::Hash> ChainFilter<H> {
 
     /// Take all actions until this timestamp is reached
     pub fn until_timestamp(mut self, timestamp: Timestamp) -> Self {
-        self.filters = match self.filters {
-            ChainFilters::ToGenesis => ChainFilters::UntilTimestamp(timestamp),
-            ChainFilters::Take(n) => ChainFilters::Multiple(n, HashSet::new(), timestamp),
-            ChainFilters::UntilTimestamp(_) => ChainFilters::UntilTimestamp(timestamp),
-            ChainFilters::UntilHash(u) => ChainFilters::Multiple(u32::MAX, u, timestamp),
-            ChainFilters::Multiple(n, u, _) => ChainFilters::Multiple(n, u, timestamp),
+        self.stop_conditions = match self.stop_conditions {
+            StopConditions::ToGenesis => StopConditions::UntilTimestamp(timestamp),
+            StopConditions::Take(n) => StopConditions::Multiple(n, HashSet::new(), timestamp),
+            StopConditions::UntilTimestamp(old_ts) => {
+                StopConditions::UntilTimestamp(timestamp.max(old_ts))
+            }
+            StopConditions::UntilHash(u) => StopConditions::Multiple(u32::MAX, u, timestamp),
+            StopConditions::Multiple(n, u, _) => StopConditions::Multiple(n, u, timestamp),
         };
         self
     }
@@ -155,23 +160,23 @@ impl<H: Eq + Ord + std::hash::Hash> ChainFilter<H> {
     /// It is invalid to specify an until hash that is on a different
     /// fork then the starting position.
     pub fn until_hash(mut self, action_hash: H) -> Self {
-        self.filters = match self.filters {
-            ChainFilters::ToGenesis => {
-                ChainFilters::UntilHash(Some(action_hash).into_iter().collect())
+        self.stop_conditions = match self.stop_conditions {
+            StopConditions::ToGenesis => {
+                StopConditions::UntilHash(Some(action_hash).into_iter().collect())
             }
-            ChainFilters::Take(n) => {
-                ChainFilters::Multiple(n, Some(action_hash).into_iter().collect(), Timestamp(0))
+            StopConditions::Take(n) => {
+                StopConditions::Multiple(n, Some(action_hash).into_iter().collect(), Timestamp(0))
             }
-            ChainFilters::UntilTimestamp(t) => {
-                ChainFilters::Multiple(u32::MAX, Some(action_hash).into_iter().collect(), t)
+            StopConditions::UntilTimestamp(t) => {
+                StopConditions::Multiple(u32::MAX, Some(action_hash).into_iter().collect(), t)
             }
-            ChainFilters::UntilHash(mut u) => {
+            StopConditions::UntilHash(mut u) => {
                 u.insert(action_hash);
-                ChainFilters::UntilHash(u)
+                StopConditions::UntilHash(u)
             }
-            ChainFilters::Multiple(n, mut u, t) => {
+            StopConditions::Multiple(n, mut u, t) => {
                 u.insert(action_hash);
-                ChainFilters::Multiple(n, u, t)
+                StopConditions::Multiple(n, u, t)
             }
         };
         self
@@ -179,18 +184,18 @@ impl<H: Eq + Ord + std::hash::Hash> ChainFilter<H> {
 
     /// Get the until hashes if there are any.
     pub fn get_until_hash(&self) -> Option<&HashSet<H>> {
-        match &self.filters {
-            ChainFilters::UntilHash(u) => Some(u),
-            ChainFilters::Multiple(_, u, _) => Some(u),
+        match &self.stop_conditions {
+            StopConditions::UntilHash(u) => Some(u),
+            StopConditions::Multiple(_, u, _) => Some(u),
             _ => None,
         }
     }
 
     /// Get the take number if there is one.
     pub fn get_take(&self) -> Option<u32> {
-        match &self.filters {
-            ChainFilters::Take(n) => Some(*n),
-            ChainFilters::Multiple(n, _, _) => {
+        match &self.stop_conditions {
+            StopConditions::Take(n) => Some(*n),
+            StopConditions::Multiple(n, _, _) => {
                 if *n != u32::MAX {
                     Some(*n)
                 } else {
@@ -203,9 +208,9 @@ impl<H: Eq + Ord + std::hash::Hash> ChainFilter<H> {
 
     /// Get the take number if there is one.
     pub fn get_until_timestamp(&self) -> Option<Timestamp> {
-        match &self.filters {
-            ChainFilters::UntilTimestamp(ts) => Some(*ts),
-            ChainFilters::Multiple(_, _, ts) => {
+        match &self.stop_conditions {
+            StopConditions::UntilTimestamp(ts) => Some(*ts),
+            StopConditions::Multiple(_, _, ts) => {
                 if ts.0 > 0 {
                     Some(*ts)
                 } else {
@@ -217,7 +222,7 @@ impl<H: Eq + Ord + std::hash::Hash> ChainFilter<H> {
     }
 }
 
-impl<H: Eq + Ord + std::hash::Hash> Default for ChainFilters<H> {
+impl<H: Eq + Ord + std::hash::Hash> Default for StopConditions<H> {
     fn default() -> Self {
         Self::ToGenesis
     }