+++ crates/holochain_integrity_types/CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+ - **BREAKING**: Renamed enum `ChainFilters` to `StopConditions` and renamed some of its variants.
+ - Adds `UntilTimestamp` variant to `ChainFilters`
+
 ## 0.6.0-dev.6
 
 ## 0.6.0-dev.5